hostel

+++ src/main/java/com/usman/hostelmanagementsystem/dto/StaffDto.java
@@ -5,13 +5,14 @@ import com.usman.hostelmanagementsystem.model.Address;
 import lombok.Data;
 
 
+import java.math.BigInteger;
 import java.util.Date;
 @Data
 public class StaffDto extends BaseDto {
     private  Long id;
     private String firstName;
     private   String surname;
-    private   Integer   identityNumber;
+    private BigInteger   identityNumber;
     private String country;
     private String  image;
     private  Integer  telefon;
